dвынесла отдельно логику по статистике
